整数から小数にした

+++ khimoo-portfolio/src/home/app.rs
@@ -7,7 +7,7 @@ use yew_hooks::{use_effect_update_with_deps, use_measure};
 #[function_component(Home)]
 pub fn home() -> Html {
     let container_ref = use_node_ref();
-    let container_pos_handle = use_state(|| Position { x: 0, y: 0 });
+    let container_pos_handle = use_state(|| Position { x: 0.0, y: 0.0 });
     let container_measure_handle = use_measure(container_ref.clone());
     let mouse_pos_handle = use_state(|| Position::default());
 
@@ -20,8 +20,8 @@ pub fn home() -> Html {
                 if let Some(element) = node_ref.cast::<HtmlElement>() {
                     let rect = element.get_bounding_client_rect();
                     container_pos.set(Position {
-                        x: rect.x() as i32,
-                        y: rect.y() as i32,
+                        x: rect.x() as f32,
+                        y: rect.y() as f32,
                     });
                 }
                 || {}
@@ -34,8 +34,8 @@ pub fn home() -> Html {
         let mouse_pos_handle = mouse_pos_handle.clone();
         Callback::from(move |e: MouseEvent| {
             let pos = Position {
-                x: e.client_x(),
-                y: e.client_y(),
+                x: e.client_x() as f32,
+                y: e.client_y() as f32,
             };
             mouse_pos_handle.set(pos);
         })
@@ -49,14 +49,14 @@ pub fn home() -> Html {
                 container_measure={container_measure_handle.clone()}
                 container_bound={
                     ContainerBound {
-                        x: container_measure_handle.x as i32 + container_pos_handle.x,
-                        y: container_measure_handle.y as i32 + container_pos_handle.y,
-                        width: container_measure_handle.width as i32,
-                        height: container_measure_handle.height as i32,
-                        top: container_measure_handle.top as i32 + container_pos_handle.y,
-                        left: container_measure_handle.left as i32 + container_pos_handle.x,
-                        bottom: container_measure_handle.bottom as i32 + container_pos_handle.y,
-                        right: container_measure_handle.right as i32 + container_pos_handle.x,
+                        x: container_measure_handle.x as f32 + container_pos_handle.x,
+                        y: container_measure_handle.y as f32 + container_pos_handle.y,
+                        width: container_measure_handle.width as f32,
+                        height: container_measure_handle.height as f32,
+                        top: container_measure_handle.top as f32 + container_pos_handle.y,
+                        left: container_measure_handle.left as f32 + container_pos_handle.x,
+                        bottom: container_measure_handle.bottom as f32 + container_pos_handle.y,
+                        right: container_measure_handle.right as f32 + container_pos_handle.x,
                     }
                 }
             />
